testgen: work in progress - add text nodes #61
Adds text nodes to the expected exile documents, which closes #61.
Deletes the C# project that did not work out, and adds a lot of work
that has been done to date on the Java project.

Does not touch the implementation in exile or the tests that are checked
in for exile, which means that the GitHub action which checks for
variance in generated test files should fail for this commit.