testgen: add processing instructions #62
Processing instructions are now added at the document level (in the
prologue and epilogue sections) and as children of elements. This gives
us the ability to generated expected functions for all of our well-
formed custom exile XML files.

Currently these files are copied, meaning 2 copies of each are checked-
in with this commit. In the future they will be moved and the copying
will stop (see #65) leaving one copy of each file.