support cdata sections
Adds support for CDATA sections with a fair amount of refactoring in the
parser. Parser functions were inconsistent about what state the iterator
would be in after a parse function. Most (maybe all) of these are now
aligned on returning from a parse function with the iterator pointing at
the first character following the construct that was parsed.